refactor storageTrace for serialize
