+++ types/src/eth.rs
@@ -1,4 +1,3 @@
-use super::mpt::StorageTrace;
 use eth_types::evm_types::{Gas, GasCost, Memory, OpcodeId, ProgramCounter, Stack, Storage};
 use eth_types::{
     fix_geth_trace_memory_size, Block, GethExecStep, GethExecTrace, Hash, Transaction, Word, H256,
@@ -22,10 +21,24 @@ pub struct BlockResult {
     pub block_trace: BlockTrace,
     #[serde(rename = "executionResults")]
     pub execution_results: Vec<ExecutionResult>,
-    #[serde(rename = "mptwitness", skip_serializing, default)]
-    pub mpt_witness: Vec<SMTTrace>,
-    #[serde(rename = "storageTrace", skip_serializing)]
+    #[serde(rename = "storageTrace")]
     pub storage_trace: StorageTrace,
+    #[serde(rename = "mptwitness", default)]
+    pub mpt_witness: Vec<SMTTrace>,
+}
+
+pub type AccountTrieProofs = HashMap::<Address, Vec<Bytes>>;
+pub type StorageTrieProofs = HashMap::<Address, HashMap<Word, Vec<Bytes>>>;
+
+#[derive(Deserialize, Serialize, Default, Debug)]
+pub struct StorageTrace {
+    #[serde(rename = "rootBefore")]
+    pub root_before: Hash,
+    #[serde(rename = "rootAfter")]
+    pub root_after: Hash,
+    pub proofs: Option<AccountTrieProofs>,
+    #[serde(rename = "storageProofs", default)]
+    pub storage_proofs: StorageTrieProofs,
 }
 
 #[derive(Deserialize, Serialize, Default, Debug, Clone)]
