+++ types/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod eth;
-pub mod mpt;
 
 pub mod base64 {
     use base64::{decode, encode};
