+++ zkevm/src/circuit.rs
@@ -16,6 +16,7 @@ use zkevm_circuits::evm_circuit::witness::{Block, RwMap};
 use zkevm_circuits::state_circuit::StateCircuitLight as StateCircuitImpl;
 
 mod builder;
+mod mpt;
 
 use crate::circuit::builder::get_fixed_table_tags_for_block;
 use crate::utils::read_env_var;
